refactor: Car테스트 Mockito 제거
