+++ build.gradle
@@ -11,7 +11,6 @@ repositories {
 dependencies {
     testCompile('org.junit.jupiter:junit-jupiter:5.4.2')
     testCompile('org.assertj:assertj-core:3.11.1')
-    testCompile('org.mockito:mockito-core:3.3.3')
 
 }
