+++ src/main/java/me/daeho/step5/README.md
@@ -1,5 +1,9 @@
 ## 5단계 - 자동차 경주
 ### 구현 기능
  - Package 구조 변경 
-   1. RacingCar, Car 객체 domain 패키지로 이동
+   1. RacingCar, Car, ForwardRule, DefaultForwardRule 객체 domain 패키지로 이동
    2. InputView, ResultView 객체 view 패키지로 이동
+   
+ - Car 테스트 
+   1. Mockito 사용하지 않고 테스트를 위해 inner Class로 테스트용 ForwardRule 구현체 사용 
+
