+++ src/main/java/me/daeho/step5/domain/ForwardRule.java
@@ -1,4 +1,4 @@
-package me.daeho.step3.rule;
+package me.daeho.step5.domain;
 
 public interface ForwardRule {
     boolean isForward();
