+++ src/main/java/me/daeho/step5/domain/RacingGame.java
@@ -1,7 +1,4 @@
-package me.daeho.step3;
-
-import me.daeho.step3.rule.DefaultForwardRule;
-import me.daeho.step3.rule.ForwardRule;
+package me.daeho.step5.domain;
 
 import java.util.*;
 import java.util.stream.Collectors;
@@ -9,9 +6,9 @@ import java.util.stream.Collectors;
 public class RacingGame {
     private int firstGradePosition;
     private int time;
-    private List<Car> cars;
+    private Collection<Car> cars;
 
-    private RacingGame(List<Car> cars, int time) {
+    private RacingGame(Collection<Car> cars, int time) {
         this.cars = cars;
         this.time = time;
     }
@@ -20,15 +17,19 @@ public class RacingGame {
         return new RacingGame(readyCars(forwardRule, carNames), time);
     }
 
-    private static List<Car> readyCars(ForwardRule forwardRule, String[] carNames) {
-        return new ArrayList<Car>(){{
+    public static RacingGame init(String[] carNames, int time) {
+        return new RacingGame(readyCars(new DefaultForwardRule(), carNames), time);
+    }
+
+    private static Collection<Car> readyCars(ForwardRule forwardRule, String[] carNames) {
+        return new ArrayList<Car>() {{
             Arrays.stream(carNames).forEach(name -> add(Car.ready(forwardRule, name)));
         }};
     }
 
-    public List<Car> move() {
+    public Collection<Car> move() {
         cars.forEach(car -> setFirstGradePosition(car.move()));
-        time --;
+        time--;
         return cars;
     }
 
@@ -36,7 +37,7 @@ public class RacingGame {
         return time > 0;
     }
 
-    public List<Car> winningCars() {
+    public Collection<Car> winningCars() {
         return cars.stream().filter(car -> car.getCurrentPosition() == firstGradePosition).collect(Collectors.toList());
     }
 
@@ -45,7 +46,7 @@ public class RacingGame {
     }
 
     private void setFirstGradePosition(int position) {
-        if(isFirstGrade(position))
+        if (isFirstGrade(position))
             firstGradePosition = position;
     }
 }
